+++ packages/site/src/componentList.ts
@@ -1,5 +1,23 @@
 export const componentList = [
   {
+    title: "ActionItem",
+    to: "/components/ActionItem",
+    imageURL: "/ActionItem.png",
+    additionalMatches: [],
+  },
+  {
+    title: "ActionItemGroup",
+    to: "/components/ActionItemGroup",
+    imageURL: "/ActionItemGroup.png",
+    additionalMatches: [],
+  },
+  {
+    title: "ActionLabel",
+    to: "/components/ActionLabel",
+    imageURL: "/ActionLabel.png",
+    additionalMatches: [],
+  },
+  {
     title: "AnimatedPresence",
     to: "/components/AnimatedPresence",
     imageURL: "/Animation.png",
@@ -15,21 +33,35 @@ export const componentList = [
     title: "Autocomplete",
     to: "/components/Autocomplete",
     imageURL: "/Autocomplete.png",
+    additionalMatches: ["Search", "Typeahead"],
+  },
+  {
+    title: "AutoLink",
+    to: "/components/AutoLink",
+    imageURL: "/AutoLink.png",
   },
   {
     title: "Avatar",
     to: "/components/Avatar",
     imageURL: "/Avatar.png",
+    additionalMatches: ["User", "Facepile", "Profile"],
   },
   {
     title: "Banner",
     to: "/components/Banner",
     imageURL: "/Banner.png",
+    additionalMatches: ["Flash", "Alert", "Notification"],
   },
   {
     title: "Box",
     to: "/components/Box",
     imageURL: "/Box.png",
+    additionalMatches: ["Layout", "Container", "Stack"],
+  },
+  {
+    title: "BottomSheet",
+    to: "/components/BottomSheet",
+    imageURL: "/BottomSheet.png",
   },
   {
     title: "Button",
@@ -41,6 +73,11 @@ export const componentList = [
     to: "/components/ButtonDismiss",
     imageURL: "/ButtonDismiss.png",
     additionalMatches: ["Close", "Dismiss", "Remove"],
+  },
+  {
+    title: "ButtonGroup",
+    to: "/components/ButtonGroup",
+    imageURL: "/ButtonGroup.png",
   },
   {
     title: "Card",
@@ -68,12 +105,24 @@ export const componentList = [
     title: "Combobox",
     to: "/components/Combobox",
     imageURL: "/Combobox.png",
+    additionalMatches: ["Dropdown", "Picker"],
   },
   {
     title: "ConfirmationModal",
     to: "/components/ConfirmationModal",
     imageURL: "/ConfirmationModal.png",
     additionalMatches: ["Confirm", "Prompt", "Modal"],
+  },
+  {
+    title: "Content",
+    to: "/components/Content",
+    imageURL: "/Content.png",
+    additionalMatches: ["Stack", "Layout"],
+  },
+  {
+    title: "ContentOverlay",
+    to: "/components/ContentOverlay",
+    imageURL: "/ContentOverlay.png",
   },
   {
     title: "Countdown",
@@ -89,21 +138,25 @@ export const componentList = [
     title: "DataList",
     to: "/components/DataList",
     imageURL: "/DataList.png",
+    additionalMatches: ["Thicklist", "Table"],
   },
   {
     title: "DataTable",
     to: "/components/DataTable",
     imageURL: "/DataTable.png",
+    additionalMatches: ["Thicklist", "Table"],
   },
   {
     title: "DatePicker",
     to: "/components/DatePicker",
     imageURL: "/DatePicker.png",
+    additionalMatches: ["Calendar", "Date Picker", "Date Input"],
   },
   {
     title: "DescriptionList",
     to: "/components/DescriptionList",
     imageURL: "/DescriptionList.png",
+    additionalMatches: ["Simple List", "Key Value Pair"],
   },
   {
     title: "Disclosure",
@@ -121,17 +174,25 @@ export const componentList = [
     title: "Divider",
     to: "/components/Divider",
     imageURL: "/Divider.png",
+    additionalMatches: ["Border", "HR"],
   },
   {
     title: "Drawer",
     to: "/components/Drawer",
     imageURL: "/Drawer.png",
+    additionalMatches: ["Aside", "Panel"],
   },
   {
     title: "Emphasis",
     to: "/components/Emphasis",
     imageURL: "/Emphasis.png",
     additionalMatches: ["Highlight", "Strong", "Bold", "Italic"],
+  },
+  {
+    title: "EmptyState",
+    to: "/components/EmptyState",
+    imageURL: "/EmptyState.png",
+    additionalMatches: ["No", "None", "Nada", "Nothing"],
   },
   {
     title: "FeatureSwitch",
@@ -142,6 +203,7 @@ export const componentList = [
     title: "Flex",
     to: "/components/Flex",
     imageURL: "/Flex.png",
+    additionalMatches: ["Layout", "Stack"],
   },
   {
     title: "Form",
@@ -182,6 +244,7 @@ export const componentList = [
     title: "Gallery",
     to: "/components/Gallery",
     imageURL: "/Gallery.png",
+    additionalMatches: ["ThumbnailList", "ImageList", "Carousel"],
   },
   {
     title: "Glimmer",
@@ -193,6 +256,7 @@ export const componentList = [
     title: "Grid",
     to: "/components/Grid",
     imageURL: "/Grid.png",
+    additionalMatches: ["Layout", "Columns"],
   },
   {
     title: "Heading",
@@ -204,6 +268,11 @@ export const componentList = [
     title: "Icon",
     to: "/components/Icon",
     imageURL: "/Icon.png",
+  },
+  {
+    title: "IconButton",
+    to: "/components/IconButton",
+    imageURL: "/IconButton.png",
   },
   {
     title: "InlineLabel",
@@ -234,6 +303,18 @@ export const componentList = [
     title: "InputEmail",
     to: "/components/InputEmail",
     imageURL: "/InputEmail.png",
+    additionalMatches: ["Email Input"],
+  },
+  {
+    title: "InputFieldWrapper",
+    to: "/components/InputFieldWrapper",
+    imageURL: "/InputFieldWrapper.png",
+  },
+  {
+    title: "InputFile",
+    to: "/components/InputFile",
+    imageURL: "/InputFile.png",
+    additionalMatches: ["Uploader", "File Input", "File Uploader"],
   },
   {
     title: "InputGroup",
@@ -244,36 +325,53 @@ export const componentList = [
     title: "InputNumber",
     to: "/components/InputNumber",
     imageURL: "/InputNumber.png",
+    additionalMatches: ["Number Input", "Numeric Input"],
   },
   {
     title: "InputPassword",
     to: "/components/InputPassword",
     imageURL: "/InputPassword.png",
+    additionalMatches: ["Password Input"],
   },
   {
     title: "InputPhoneNumber",
     to: "/components/InputPhoneNumber",
     imageURL: "/InputPhoneNumber.png",
+    additionalMatches: ["Phone Number Input"],
+  },
+  {
+    title: "InputPressable",
+    to: "/components/InputPressable",
+    imageURL: "/InputPressable.png",
+  },
+  {
+    title: "InputSearch",
+    to: "/components/InputSearch",
+    imageURL: "/InputSearch.png",
   },
   {
     title: "InputText",
     to: "/components/InputText",
     imageURL: "/InputText.png",
+    additionalMatches: ["Textfield", "Text Input", "Input"],
   },
   {
     title: "InputTime",
     to: "/components/InputTime",
     imageURL: "/InputTime.png",
+    additionalMatches: ["Time Input", "Time Picker"],
   },
   {
     title: "InputValidation",
     to: "/components/InputValidation",
     imageURL: "/InputValidation.png",
+    additionalMatches: ["Field Error", "Error Message"],
   },
   {
     title: "LightBox",
     to: "/components/LightBox",
     imageURL: "/LightBox.png",
+    additionalMatches: ["Gallery", "Carousel"],
   },
   {
     title: "Link",
@@ -295,53 +393,79 @@ export const componentList = [
     title: "Menu",
     to: "/components/Menu",
     imageURL: "/Menu.png",
+    additionalMatches: ["Dropdown"],
   },
   {
     title: "Modal",
     to: "/components/Modal",
     imageURL: "/Modal.png",
+    additionalMatches: ["Dialog", "Popup", "Overlay"],
   },
   {
     title: "MultiSelect",
     to: "/components/MultiSelect",
     imageURL: "/MultiSelect.png",
+    additionalMatches: ["Dropdown"],
+  },
+  {
+    title: "Page",
+    to: "/components/Page",
+    imageURL: "/Page.png",
   },
   {
     title: "Popover",
     to: "/components/Popover",
     imageURL: "/Popover.png",
+    additionalMatches: ["Dropdown", "Popup", "Tooltip"],
   },
   {
     title: "ProgressBar",
     to: "/components/ProgressBar",
     imageURL: "/ProgressBar.png",
-    additionalMatches: ["Progress", "Loader"],
+    additionalMatches: ["Progress", "Loader", "Stepper"],
   },
   {
     title: "RadioGroup",
     to: "/components/RadioGroup",
     imageURL: "/RadioGroup.png",
+    additionalMatches: ["Radio Button"],
   },
   {
-    title: "ReccuringSelect",
+    title: "RecurringSelect",
     to: "/components/RecurringSelect",
     imageURL: "/RecurringSelect.png",
+    additionalMatches: ["Calendar", "Date Picker", "Schedule"],
   },
   {
     title: "SegmentedControl",
     to: "/components/SegmentedControl",
     imageURL: "/SegmentedControl.png",
+    additionalMatches: ["ViewControl", "TabControl", "Tab"],
   },
   {
     title: "Select",
     to: "/components/Select",
     imageURL: "/Select.png",
+    additionalMatches: ["Dropdown", "Picker"],
+  },
+  {
+    title: "SideDrawer",
+    to: "/components/SideDrawer",
+    imageURL: "/SideDrawer.png",
+    additionalMatches: ["Drawer", "Panel", "Sidebar"],
   },
   {
     title: "Spinner",
     to: "/components/Spinner",
     imageURL: "/Spinner.png",
-    additionalMatches: ["Loader", "Loading", "Twirly", "Wait", "Busy"],
+    additionalMatches: [
+      "Loader",
+      "Loading",
+      "Twirly",
+      "Wait",
+      "Busy",
+      "ActivityIndicator",
+    ],
   },
   {
     title: "StatusIndicator",
@@ -369,21 +493,35 @@ export const componentList = [
     title: "Tabs",
     to: "/components/Tabs",
     imageURL: "/Tabs.png",
+    additionalMatches: ["Segmented Control"],
   },
   {
     title: "Text",
     to: "/components/Text",
     imageURL: "/Text.png",
+    additionalMatches: ["Paragraph", "Body Copy"],
+  },
+  {
+    title: "TextList",
+    to: "/components/TextList",
+    imageURL: "/TextList.png",
+  },
+  {
+    title: "ThumbnailList",
+    to: "/components/ThumbnailList",
+    imageURL: "/ThumbnailList.png",
   },
   {
     title: "Toast",
     to: "/components/Toast",
     imageURL: "/Toast.png",
+    additionalMatches: ["Alert", "Notification", "Flash"],
   },
   {
     title: "Tooltip",
     to: "/components/Tooltip",
     imageURL: "/Tooltip.png",
+    additionalMatches: ["Popup", "Hover", "Popover"],
   },
   {
     title: "Typography",