+++ packages/site/src/content/Checkbox/index.tsx
@@ -1,11 +1,13 @@
 import CheckboxContent from "@atlantis/docs/components/Checkbox/Checkbox.stories.mdx";
 import Props from "./Checkbox.props.json";
+import MobileProps from "./Checkbox.props-mobile.json";
 import { ContentExport } from "../../types/content";
 import { getStorybookUrl } from "../../layout/getStorybookUrl";
 
 export default {
   content: () => <CheckboxContent />,
   props: Props,
+  mobileProps: MobileProps,
   component: {
     element: `const [checked, setChecked] = useState(true);
 
@@ -16,6 +18,7 @@ export default {
       onChange={setChecked}
     />
   );`,
+    mobileElement: `<Checkbox label={"Check me out"} name={"storyCheckbox"} />`,
   },
   title: "Checkbox",
   links: [