+++ packages/site/src/content/Typography/index.tsx
@@ -1,13 +1,15 @@
 import Content from "@atlantis/docs/components/Typography/Typography.stories.mdx";
 import Props from "./Typography.props.json";
+import MobileProps from "./Typography.props-mobile.json";
 import { ContentExport } from "../../types/content";
 
 export default {
   content: () => <Content />,
   props: Props,
+  mobileProps: MobileProps,
   component: {
     element: `<Typography size='large'>Typography</Typography>`,
-    defaultProps: {},
+    mobileElement: `<Typography size='large'>Typography</Typography>`,
   },
   title: "Typography",
   links: [