+++ packages/site/src/layout/Layout.tsx
@@ -3,6 +3,7 @@ import { Route, Switch, useLocation } from "react-router";
 import { NavMenu } from "./NavMenu";
 import { routes } from "../routes";
 import "./code-theme.css";
+import { ToggleThemeButton } from "../components/ToggleThemeButton";
 
 /**
  * Layout for whole application. This will display the NavMenu and the content of the page.
@@ -12,10 +13,7 @@ export const Layout = () => {
   const location = useLocation();
 
   useEffect(() => {
-    // Perform an action whenever the location changes
     window.scrollTo({ top: 0 });
-
-    // Add your custom logic here
   }, [location]);
 
   return (
@@ -44,6 +42,8 @@ export const Layout = () => {
           )}
         </Switch>
       </div>
+
+      <ToggleThemeButton />
     </LayoutWrapper>
   );
 };