+++ packages/site/src/layout/NavMenu.tsx
@@ -1,9 +1,10 @@
-import { Box, Button } from "@jobber/components";
+import { Box, Button, Disclosure, Typography } from "@jobber/components";
 import { Link } from "react-router-dom";
 import { PropsWithChildren, useState } from "react";
 import { SearchBox } from "./SearchBox";
 import { routes } from "../routes";
 import { JobberLogo } from "../assets/JobberLogo.svg";
+import { useAtlantisSite } from "../providers/AtlantisSiteProvider";
 
 /**
  * Left side navigation menu for the application.
@@ -11,13 +12,17 @@ import { JobberLogo } from "../assets/JobberLogo.svg";
  */
 export const NavMenu = () => {
   const [open, setOpen] = useState(false);
+  const { isMinimal } = useAtlantisSite();
+
+  if (isMinimal) return null;
 
   return (
     <div
       style={{
         width: 220,
-        minHeight: "100dvh",
+        height: "100dvh",
         backgroundColor: "var(--color-surface--background)",
+        overflow: "scroll",
       }}
     >
       <Box height={24} padding="base">
@@ -37,19 +42,28 @@ export const NavMenu = () => {
       <MenuList>
         <Box>
           {routes?.map((route, index) => {
-            return route.children ? (
-              route.children.map((subroute, subindex) => {
-                if (subroute.inNav === false) return;
+            if (route.children) {
+              return (
+                <Box key={index} padding="base">
+                  <Disclosure key={index} title={changelogTitle}>
+                    {route.children.map((subroute, subindex) => {
+                      if (subroute.inNav === false) return null;
 
-                return (
-                  <MenuItem key={index}>
-                    <StyledLink to={subroute.path || "/"} key={subindex}>
-                      {subroute.handle}
-                    </StyledLink>
-                  </MenuItem>
-                );
-              })
-            ) : route.inNav === false ? null : (
+                      return (
+                        <div key={subindex}>
+                          <StyledSubLink to={subroute.path || "/"}>
+                            {subroute.handle}
+                          </StyledSubLink>
+                        </div>
+                      );
+                    })}
+                  </Disclosure>
+                </Box>
+              );
+            }
+            if (route.inNav === false) return null;
+
+            return (
               <MenuItem key={index}>
                 <StyledLink
                   key={index}
@@ -93,6 +107,31 @@ export const StyledLink = ({
   );
 };
 
+export const StyledSubLink = ({
+  to,
+  children,
+  style,
+}: PropsWithChildren<{ readonly to: string; readonly style?: object }>) => {
+  return (
+    <Link
+      to={to ?? "/"}
+      style={{
+        outline: "transparent",
+        color: "var(--color-heading)",
+        fontSize: "var(--typography--fontSize-base)",
+        fontWeight: 700,
+        width: "100%",
+        textDecoration: "none",
+        userSelect: "none",
+        transition: "all var(--timing-base) ease-out",
+        ...style,
+      }}
+    >
+      {children}
+    </Link>
+  );
+};
+
 export const MenuList = ({ children }: PropsWithChildren) => {
   return <ul style={{ listStyleType: "none", padding: 0 }}>{children}</ul>;
 };
@@ -113,3 +152,9 @@ export const MenuItem = ({ children }: PropsWithChildren) => {
     </li>
   );
 };
+
+export const changelogTitle = (
+  <Typography fontWeight="semiBold" size="large" textColor="heading">
+    Changelog
+  </Typography>
+);