+++ packages/site/src/pages/ComponentsPage.tsx
@@ -6,8 +6,9 @@ export const ComponentsPage = () => {
     <PageBlock
       structure={{
         header: {
-          title: "Web components",
-          body: "They are the best",
+          title: "Components",
+          body: "The tools you'll use to build Jobber",
+          imageURL: "../public/img-hero_collage-v2.004b0168.webp",
         },
         body: {
           title: "Components",